(#97) moving DTOs to the domain
