+++ storage/src/main/java/org/zalando/apidiscovery/storage/api/resource/ApplicationResourceController.java
@@ -9,8 +9,8 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.util.UriComponentsBuilder;
-import org.zalando.apidiscovery.storage.api.service.dto.ApplicationDto;
-import org.zalando.apidiscovery.storage.api.service.dto.ApplicationsDto;
+import org.zalando.apidiscovery.storage.api.domain.dto.ApplicationDto;
+import org.zalando.apidiscovery.storage.api.domain.dto.ApplicationsDto;
 import org.zalando.apidiscovery.storage.api.service.ApplicationService;
 
 import java.util.List;
