+++ storage/src/test/java/org/zalando/apidiscovery/storage/api/domain/dto/ApiDtoJsonTest.java
@@ -1,4 +1,4 @@
-package org.zalando.apidiscovery.storage.api.service.dto;
+package org.zalando.apidiscovery.storage.api.domain.dto;
 
 
 import org.junit.Test;
